fixes and outputs

+++ utils/TimeUtils.h
@@ -21,7 +21,7 @@ public:
         return ss.str();
     }
 
-    static std::string FormatMicrosecondToMilliseconds(long microsec)
+    static std::string FormatMicrosecondToMilliseconds(long long microsec)
     {
         std::stringstream ss;
         ss << std::fixed << std::setprecision(3) << microsec / 1000.0;
